CEXIO: Fix buy maker size calculation
Closes already closed #923

CEXIO (at least) always expects the taker fee will submitting a buy order which causes a ìnsufficient funds` error if buy_pct (+ makerFee) is greater than or equal to 100% - takerFee

To workaround this "limitation" we will use the taker fee for the buy order size calculation.

Note that this has no impact on the actual fee paid if the order gets filled, its just to get the order accepted.

This is an exchange specific workaround, so if you suspect that other exchanges have the same "problem" try adding `makerBuy100Workaround: true` to the exchange object in exchange.js